Refactor Entity Numbers Structure
Remove individual interfaces for user account number and politician
number and create an abstract class that implements the EntityNumberInterface
