--- src/main/java/com/example/demo/model/AbstractPoliticianNumber.java
@@ -1,22 +1,0 @@
-package com.example.demo.model;
-
-public abstract class AbstractPoliticianNumber implements PoliticianNumberInterface {
-
-	protected final String firstName, lastName;
-
-	public String getFirstName() {
-		return firstName;
-	}
-
-	public String getLastName() {
-		return lastName;
-	}
-
-
-	public AbstractPoliticianNumber(String firstName, String lastName) {
-		super();
-		this.firstName = firstName;
-		this.lastName = lastName;
-	}
-
-}
