--- src/main/java/com/example/demo/model/NumberInterface.java
@@ -1,7 +1,0 @@
-package com.example.demo.model;
-
-public interface NumberInterface {
-	
-	
-	
-}
