--- src/main/java/com/example/demo/model/PoliticianNumberInterface.java
@@ -1,6 +1,0 @@
-package com.example.demo.model;
-
-public interface PoliticianNumberInterface extends NumberInterface{
-
-	PoliticianNumberInterface calculatePoliticianNumber();
-}
