--- src/main/java/com/example/demo/model/UserRaterNumberInterface.java
@@ -1,9 +1,0 @@
-package com.example.demo.model;
-
-public interface UserRaterNumberInterface extends NumberInterface {
-
-	UserRaterNumberInterface calculateUserRaterAccountNumber();
-	
-	String returnAccountNumber();
-	
-}
