+++ src/main/java/com/example/demo/model/entityNumber/NumberInterface.java
@@ -1,0 +1,7 @@
+package com.example.demo.model.entityNumber;
+
+public interface NumberInterface {
+	
+	NumberInterface calculateEntityNumber();
+	
+}
